refactor(http): Update paths for explorers
